// examples/client/src/gen7/models/ts/pet/AddPet.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { Pet } from '../Pet.ts'
import type { PetNotFound } from '../PetNotFound.ts'

/**
 * @type object
 */
export type AddPetStatus200Json = Pet

/**
 * @type object
 */
export type AddPetStatus200Xml = Pet

export type AddPetStatus200 = AddPetStatus200Json | AddPetStatus200Xml

export type AddPetStatus405 = PetNotFound

/**
 * @description Create a new pet in the store
 * @type object
 */
export type AddPetJsonData = Pet

/**
 * @description Create a new pet in the store
 * @type object
 */
export type AddPetXmlData = Pet

/**
 * @description Create a new pet in the store
 * @type object
 */
export type AddPetFormUrlEncodedData = Pet

export type AddPetData = AddPetJsonData | AddPetXmlData | AddPetFormUrlEncodedData

/**
 * @type object
 */
export type AddPetRequestConfig = {
  data?: AddPetData
  pathParams?: never
  queryParams?: never
  headerParams?: never
  /**
   * @type string
   */
  url: '/pet'
}

/**
 * @type object
 */
export type AddPetResponses = {
  '200': AddPetStatus200
  '405': AddPetStatus405
}

/**
 * @description Union of all possible responses
 */
export type AddPetResponse = AddPetStatus200 | AddPetStatus405