// examples/client/src/gen7/models/ts/pet/DeletePet.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

/**
 * @type string | undefined
 */
export type DeletePetHeaderApiKey = string | undefined

/**
 * @description Pet id to delete
 *
 * Format: `int64`
 * @type integer
 */
export type DeletePetPathPetId = bigint

/**
 * @type any
 */
export type DeletePetStatus400 = any

/**
 * @type object
 */
export type DeletePetRequestConfig = {
  data?: never
  /**
   * @type object
   */
  pathParams: {
    petId: DeletePetPathPetId
  }
  queryParams?: never
  /**
   * @type object | undefined
   */
  headerParams?: {
    api_key?: DeletePetHeaderApiKey
  }
  /**
   * @type string
   */
  url: `/pet/${string}`
}

/**
 * @type object
 */
export type DeletePetResponses = {
  '400': DeletePetStatus400
}

/**
 * @description Union of all possible responses
 */
export type DeletePetResponse = DeletePetStatus400