// examples/client/src/gen7/models/ts/pet/FindPetsByStatus.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { FindPetsByStatusStatusKey } from '../FindPetsByStatusStatus.ts'
import type { Pet } from '../Pet.ts'

/**
 * @description Status values that need to be considered for filter
 * @default available
 */
export type FindPetsByStatusQueryStatus = FindPetsByStatusStatusKey | undefined

/**
 * @type array
 */
export type FindPetsByStatusStatus200Json = Array<Pet>

/**
 * @type array
 */
export type FindPetsByStatusStatus200Xml = Array<Pet>

export type FindPetsByStatusStatus200 = FindPetsByStatusStatus200Json | FindPetsByStatusStatus200Xml

/**
 * @type any
 */
export type FindPetsByStatusStatus400 = any

/**
 * @type object
 */
export type FindPetsByStatusRequestConfig = {
  body?: never
  path?: never
  /**
   * @type object | undefined
   */
  query?: {
    status?: FindPetsByStatusQueryStatus
  }
  headers?: never
  /**
   * @type string
   */
  url: '/pet/findByStatus'
}

/**
 * @type object
 */
export type FindPetsByStatusResponses = {
  '200': FindPetsByStatusStatus200
  '400': FindPetsByStatusStatus400
}

/**
 * @description Union of all possible responses
 */
export type FindPetsByStatusResponse = FindPetsByStatusStatus200 | FindPetsByStatusStatus400